ref(blog/pagination): types moved, styles updated
