+++ src/styles/tokens.scss
@@ -131,8 +131,10 @@ body {
   --space-01: 0.1rem;
   --space-02: 0.2rem;
   --space-04: 0.4rem;
+  --space-07: 0.7rem;
   --space-08: 0.8rem;
   --space-12: 1.2rem;
+  --space-13: 1.3rem;
   --space-14: 1.4rem;
   --space-16: 1.6rem;
   --space-20: 2rem;
